[HTML] Fixed bug in referencing entity definitions.
In HTML5, the link to an entity, which is given as the href attribute
of an anchor tag, may only contain unreserved characters (see RFC 3986).
All other characters have to be "percent"-encoded. However, this is not
true for the id attribute definining the target of the link, which may
contain arbitrary (non-whitespace) characters. In consequence, the name
of an entity (function/constructor) must be encoded for the link, but
must not be encoded for the definition.