Add missing pattern in `renameLiteral`
Due to a change of the representation of `Float` in src/Curry/Syntax/Type, which is now handled similar to `Int`, float values are stored in the environment as well. Unfortunately, all float values were named `_` due to the missing pattern in `renameLiteral` that extends the given string with a random number at compile-time. The float value named `_` was then added into the interface files and caused an `unexpected token` error at runtime.