Desugaring of patterns corrected, fixes #1226
Nonlinear patterns have to be desugared first to eliminate multiple occurrences
of the same variable outside the scope of a single functional pattern.
This commit changes the order of pattern desugaring in that desugaring of
nonlinear patterns is now performed first.
Consequently, functional patterns have to be considered when eliminating
nonlinear patterns and the implementation is adapted accordingly.