AbstractCurry: Qualify global declarations for easier meta-programming
This change also qualifies the names in global (top-level) declarations
although the names can only be notated unqualified in Curry. This is to
faciliate the meta-programming, for instance the retrieval of a function
declaration. In addition, primitive entities such as tuples, unit and
list constructors are qualified for the same reason.